refactor: switch Plugin to direct self-reexport imports
