+++ packages/opencode/src/cli/cmd/providers.ts
@@ -9,7 +9,7 @@ import path from "path"
 import os from "os"
 import { Config } from "../../config"
 import { Global } from "../../global"
-import { Plugin } from "../../plugin"
+import { Plugin } from "../../plugin/plugin"
 import { Instance } from "../../project/instance"
 import type { Hooks } from "@opencode-ai/plugin"
 import { Process } from "../../util"
