+++ packages/opencode/src/effect/app-runtime.ts
@@ -14,7 +14,7 @@ import { File } from "@/file"
 import { FileWatcher } from "@/file/watcher"
 import { Storage } from "@/storage"
 import { Snapshot } from "@/snapshot"
-import { Plugin } from "@/plugin"
+import { Plugin } from "@/plugin/plugin"
 import { Provider } from "@/provider"
 import { ProviderAuth } from "@/provider"
 import { Agent } from "@/agent/agent"
