--- packages/opencode/src/plugin/index.ts
@@ -1,1 +1,0 @@
-export * as Plugin from "./plugin"
