+++ packages/opencode/src/plugin/plugin.ts
@@ -285,3 +285,5 @@ export const layer = Layer.effect(
 )
 
 export const defaultLayer = layer.pipe(Layer.provide(Bus.layer), Layer.provide(Config.defaultLayer))
+
+export * as Plugin from "./plugin"
