+++ packages/opencode/src/provider/auth.ts
@@ -4,7 +4,7 @@ import { Auth } from "@/auth"
 import { InstanceState } from "@/effect"
 import { zod } from "@/util/effect-zod"
 import { withStatics } from "@/util/schema"
-import { Plugin } from "../plugin"
+import { Plugin } from "../plugin/plugin"
 import { ProviderID } from "./schema"
 import { Array as Arr, Effect, Layer, Record, Result, Context, Schema } from "effect"
 import z from "zod"
