+++ packages/opencode/src/provider/provider.ts
@@ -7,7 +7,7 @@ import { NoSuchModelError, type Provider as SDK } from "ai"
 import { Log } from "../util"
 import { Npm } from "../npm"
 import { Hash } from "@opencode-ai/shared/util/hash"
-import { Plugin } from "../plugin"
+import { Plugin } from "../plugin/plugin"
 import { NamedError } from "@opencode-ai/shared/util/error"
 import { type LanguageModelV3 } from "@ai-sdk/provider"
 import * as ModelsDev from "./models"
