+++ packages/opencode/src/pty/service.ts
@@ -7,7 +7,7 @@ import z from "zod"
 import { Log } from "../util"
 import { lazy } from "@opencode-ai/shared/util/lazy"
 import { Shell } from "@/shell/shell"
-import { Plugin } from "@/plugin"
+import { Plugin } from "@/plugin/plugin"
 import { PtyID } from "./schema"
 import { Effect, Layer, Context } from "effect"
 import { EffectBridge } from "@/effect"
