+++ packages/opencode/src/session/compaction.ts
@@ -9,7 +9,7 @@ import { Token } from "../util"
 import { Log } from "../util"
 import { SessionProcessor } from "./processor"
 import { Agent } from "@/agent/agent"
-import { Plugin } from "@/plugin"
+import { Plugin } from "@/plugin/plugin"
 import { Config } from "@/config"
 import { NotFoundError } from "@/storage"
 import { ModelID, ProviderID } from "@/provider/schema"
