+++ packages/opencode/src/session/processor.ts
@@ -4,7 +4,7 @@ import { Agent } from "@/agent/agent"
 import { Bus } from "@/bus"
 import { Config } from "@/config"
 import { Permission } from "@/permission"
-import { Plugin } from "@/plugin"
+import { Plugin } from "@/plugin/plugin"
 import { Snapshot } from "@/snapshot"
 import * as Session from "./session"
 import { LLM } from "./llm"
