+++ packages/opencode/src/session/prompt.ts
@@ -15,7 +15,7 @@ import { Bus } from "../bus"
 import { ProviderTransform } from "../provider"
 import { SystemPrompt } from "./system"
 import { Instruction } from "./instruction"
-import { Plugin } from "../plugin"
+import { Plugin } from "../plugin/plugin"
 import PROMPT_PLAN from "../session/prompt/plan.txt"
 import BUILD_SWITCH from "../session/prompt/build-switch.txt"
 import MAX_STEPS from "../session/prompt/max-steps.txt"
