+++ packages/opencode/src/tool/bash.ts
@@ -16,7 +16,7 @@ import { Shell } from "@/shell/shell"
 
 import { BashArity } from "@/permission/arity"
 import * as Truncate from "./truncate"
-import { Plugin } from "@/plugin"
+import { Plugin } from "@/plugin/plugin"
 import { Effect, Stream } from "effect"
 import { ChildProcess } from "effect/unstable/process"
 import { ChildProcessSpawner } from "effect/unstable/process/ChildProcessSpawner"
