+++ packages/opencode/test/session/compaction.test.ts
@@ -12,7 +12,7 @@ import { Token } from "../../src/util"
 import { Instance } from "../../src/project/instance"
 import { Log } from "../../src/util"
 import { Permission } from "../../src/permission"
-import { Plugin } from "../../src/plugin"
+import { Plugin } from "../../src/plugin/plugin"
 import { provideTmpdirInstance, tmpdir } from "../fixture/fixture"
 import { Session as SessionNs } from "../../src/session"
 import { MessageV2 } from "../../src/session/message-v2"
