+++ packages/opencode/test/session/processor-effect.test.ts
@@ -7,7 +7,7 @@ import { Agent as AgentSvc } from "../../src/agent/agent"
 import { Bus } from "../../src/bus"
 import { Config } from "../../src/config"
 import { Permission } from "../../src/permission"
-import { Plugin } from "../../src/plugin"
+import { Plugin } from "../../src/plugin/plugin"
 import { Provider } from "../../src/provider"
 import { ModelID, ProviderID } from "../../src/provider/schema"
 import { Session } from "../../src/session"
