+++ packages/opencode/test/tool/bash.test.ts
@@ -13,7 +13,7 @@ import { Truncate } from "../../src/tool"
 import { SessionID, MessageID } from "../../src/session/schema"
 import * as CrossSpawnSpawner from "../../src/effect/cross-spawn-spawner"
 import { AppFileSystem } from "@opencode-ai/shared/filesystem"
-import { Plugin } from "../../src/plugin"
+import { Plugin } from "../../src/plugin/plugin"
 
 const runtime = ManagedRuntime.make(
   Layer.mergeAll(
